Fine tuning of grammar
Adjusted the grammar of sentence for readability, and on the odd chance that this was a secret challenge testing the editorial skills of Tick Tack and SSCB supporters, in order to pick one lucky supporter to join your intergalactic legion of decentralizers!Great project! Looking forward to learning more about it.